+++ basiccheck/slice_test.go
@@ -7,24 +7,6 @@ import (
 	"github.com/jeremmfr/go-utils/basiccheck"
 )
 
-func TestInSlice(t *testing.T) {
-	sliceOfString := []string{"foo", "bar"}
-	if !basiccheck.InSlice("bar", sliceOfString) {
-		t.Errorf("InSlice didn't find bar in %v", sliceOfString)
-	}
-	if basiccheck.InSlice("baz", sliceOfString) {
-		t.Errorf("InSlice found baz in %v", sliceOfString)
-	}
-
-	sliceOfInt := []int{2, 4, 10}
-	if !basiccheck.InSlice(4, sliceOfInt) {
-		t.Errorf("InSlice didn't find 4 in %v", sliceOfInt)
-	}
-	if basiccheck.InSlice(6, sliceOfInt) {
-		t.Errorf("InSlice found 6 in %v", sliceOfInt)
-	}
-}
-
 func TestSimilarSlice(t *testing.T) {
 	sliceA := []string{"foo", "bar", "baz"}
 	sliceB := []string{"foo", "baz", "bar"}